refactor(firmware): versat code to separated file - this prepares code to be reused in versat verilog source generation   program
